Turn warning into lint
